Revert "Frontend updates"
